tests/wpt/web-platform-tests/css/css-flexbox-1/getcomputedstyle/flexbox_computedstyle_flex-shorthand.html

<!DOCTYPE html>
<title>flexbox | computed style | flex: invalid</title>
<link rel="author" href="http://opera.com" title="Opera Software">
<link rel="help" href="http://www.w3.org/TR/css-flexbox-1/#propdef-flex">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<meta name="flags" content="dom">
<style>
body {
	color: red;
	flex: 666 666 666px;
}
</style>
<div id="log"></div>
<script>
test(function() {
    var body = document.body;

    assert_equals(getComputedStyle(body).getPropertyValue("flex-grow"),
	          "666");
    assert_equals(getComputedStyle(body).getPropertyValue("flex-shrink"),
	          "666");
    assert_equals(getComputedStyle(body).getPropertyValue("flex-basis"),
	          "666px");
});
</script>
